Fixes in serialization tests for LoggerTakingImplicit
