+++ src/test/scala/com/typesafe/scalalogging/LoggerTakingImplicitSpec.scala
@@ -1,12 +1,10 @@
 package com.typesafe.scalalogging
 
-import java.io.{ByteArrayInputStream, ByteArrayOutputStream, ObjectInputStream, ObjectOutputStream}
-
 import org.mockito.Matchers._
 import org.mockito.Mockito._
 import org.scalatest.mockito.MockitoSugar
-import org.scalatest.{Matchers, WordSpec}
-import org.slf4j.{Logger => Underlying}
+import org.scalatest.{ Matchers, WordSpec }
+import org.slf4j.{ Logger => Underlying }
 
 class LoggerTakingImplicitSpec extends WordSpec with Matchers with MockitoSugar {
 
@@ -377,69 +375,10 @@ class LoggerTakingImplicitSpec extends WordSpec with Matchers with MockitoSugar 
     }
   }
 
-  "Serializing Logger" should {
-    implicit val correlationId = CorrelationId(value = "correlationId")
-
-    implicit case object canLogCorrelationId extends CanLog[CorrelationId] {
-      override def logMessage(originalMsg: String, a: CorrelationId): String = s"${a.value} msg"
-    }
-
-    def serialize[A](logger: LoggerTakingImplicit[A]): Array[Byte] = {
-      val byteArrayStream = new ByteArrayOutputStream
-      val out = new ObjectOutputStream(byteArrayStream)
-
-      out.writeObject(logger)
-      out.close()
-      byteArrayStream.close()
-
-      byteArrayStream.toByteArray
-    }
-
-    def deserialize[A](array: Array[Byte]): LoggerTakingImplicit[A] = {
-      val byteArrayStream = new ByteArrayInputStream(array)
-      val in = new ObjectInputStream(byteArrayStream)
-
-      val logger = in.readObject.asInstanceOf[LoggerTakingImplicit[A]]
-      in.close()
-      byteArrayStream.close()
-
-      logger
-    }
-
-    "be usable after deserialization" in {
-      val logger =
-        deserialize[CorrelationId](
-          serialize[CorrelationId](
-            Logger.takingImplicit[CorrelationId](
-              org.slf4j.LoggerFactory.getLogger("test"))))
-
-      logger.trace("Back from deserialization")
-    }
-
-    "constructed by explicit class and be usable after deserialization" in {
-      val logger =
-        deserialize[CorrelationId](
-          serialize[CorrelationId](
-            Logger.takingImplicit[CorrelationId](
-              classOf[LoggerSpec])))
-
-      logger.trace("Back from deserialization")
-    }
-
-    "constructed by implicit class tag and be usable after deserialization" in {
-      val logger =
-        deserialize[CorrelationId](
-          serialize[CorrelationId](
-            Logger.takingImplicit[LoggerSpec, CorrelationId]))
-
-      logger.trace("Back from deserialization")
-    }
-  }
-
   def fixture(p: Underlying => Boolean, isEnabled: Boolean) =
     new {
       implicit val correlationId = CorrelationId("corrId")
-      implicit val canLogCorrelationId =  mock[CanLog[CorrelationId]]
+      implicit val canLogCorrelationId = mock[CanLog[CorrelationId]]
       val msg = "msg"
       val cause = new RuntimeException("cause")
       val arg1 = "arg1"
